index: Include Utils (#2)
This includes the Utils module in muzak/index.rb so that the `error`
functions are defined. This caused a bug where causing an error in
`albums-by-artist` or `songs-by-artist` would cause the program to quit.

This also returns an empty Hash instead of an empty Array when no artist is
found for `albums-by-artist`, so that it does not fail when the keys
method is called on it.