config: Remove config-set command.
Allowing the user to change the config inside of muzak itself
would cause all kinds of reloading and type-checking pain,
so remove this until a much nicer solution than the (now previous)
one is available. "config-get" is still available, since it's
harmless and useful for debugging.

index: Automatically re-build the index if outdated.

Checks for "index-autobuild" in the config, and calls index_build
if its value (a number) is less than the age of the index (from
the timestamp key). By default, "index-autobuild" is set to 86400
(1 day).

plugin: Detect plugins stored under the config tree.

Users can add plugins to ~/.config/muzak/plugins/, and they'll be loaded
like the "standard" plugins under lib/muzak/plugin. Like the plugins
bundled with muzak, user plugins are only instantiated if "plugin-<name>"
is a key in the config at startup, where "<name>" is usually the lowercased
version of the plugin's classname.