lib: Add optional deep index support.
When 'deep-index' is enabled in the configuration, muzak will attempt
to build a full array of Song objects in conjunction with the standard array
of paths for each album in the index. This takes much longer, but can be
used to amortize the cost of large one-off operations (large jukeboxes,
large playlists, etc).

Deep index data isn't currently exposed, so no performance improvements
are immediately available.