lib: Make all playlists load, instead of just default-playlist.
This is a lot more flexible than the previous system, as the user
can now add songs to any playlist without having to `playlist-load` it
first. It also allows playlists to be created on demand through any
command that accesses the playlists hash.

Since all playlists are loaded at the same time, the playlist_loaded
event had been changed to playlists_loaded.