lib: Remove index building from core. bin: Add muzak-index for index generation. COMMANDS: Remove index-build. CONFIGURATION: Remove index-autobuild and deep-index.
Building muzak's index within a muzak index itself was a sore point
in terms of performance and usability.

One forseeable use case of muzak is with a networked drive, in which
case the user will most likely want to perform the I/O intensive indexing
on the file server rather than on the machine running muzak.

Although the new index-build program still requires muzak to be installed,
it attempts to use as little of muzak's footprint as possible. This should
make it relatively easy to install a "skeleton" copy of muzak on the server
while keeping one's real configuration on the playback machine.