lib: Add player sanity checking.
`Muzak::Player::StubPlayer` now exposes an `available?` class method,
which descendants should use to check whether their components are
actually available on the user's system. This goes well with the new
`Muzak::Utils.which?` method, which can be used to test for the presence
of a utility in the user's `$PATH` (a la which(1)).

This commit also fixes `Muzak::Index#deep?` and makes `Muzak::Utils#error`
and `Muzak::Utils.debug` behave more nicely when called within a class
scope.