lib: Move constants in const.rb to Config namespace.
Finally gets rid of this ugly file. `VERSION` is preserved
in the `Muzak` namespace, for discoverability reasons.

This commit also constant-ifies some stuff in `Utils`
(like the array/regexp used to find music and art), and moves
`Utils.resolve_command` and `Utils.resolve_method` to `Config`.

It also replaces Utils.warn with Utils.danger, since `Kernel#warn` is
a thing.