player: Add MPD.
This adds support for MPD as a "player," which really just means that
it manages the MPD active playlist (and MPD does whatever it wants with
respect to its own outputs).

This is very WIP, and is probably full of bugs. It also doesn't support
events, and requires that Muzak and MPD share the same library prefix.